Extract methods to their own files
This is in preparation to check-out those files in master
before changing existing code.
