+++ pkg/controller/servicebindingrequest/plan_decoupled.go
@@ -1,0 +1,13 @@
+package servicebindingrequest
+
+import "k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+
+// GetCRs returns a slice of unstructured CRs contained in the internal related resources collection.
+func (p *Plan) GetCRs() []*unstructured.Unstructured {
+	return p.RelatedResources.GetCRs()
+}
+
+// GetRelatedResources returns the collection of related resources enumerated in the plan.
+func (p *Plan) GetRelatedResources() RelatedResources {
+	return p.RelatedResources
+}
